requirements.txt
Django==3.0.2
django-environ==0.4.5
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4